Roommate Matching App/requirements.txt
Authlib==1.3.0
bidict==0.23.1
blinker==1.7.0
boto3==1.34.85
botocore==1.34.85
cachelib==0.10.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.0
cryptography==42.0.5
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==24.14.0
flash==1.0.3
Flask==2.3.1
Flask-Cors==3.0.10
Flask-OAuth==0.12
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jwt==1.3.1
Mako==1.3.3
Markdown==3.6
MarkupSafe==2.1.3
numpy==1.26.4
oauth2==1.9.0.post1
oauthlib==2.1.0
packaging==23.2
pdoc3==0.10.0
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
requests-oauthlib==1.1.0
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
threadpoolctl==3.4.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
URLObject==2.4.3
waitress==2.1.2
Werkzeug==2.3.0
wsproto==1.2.0
